Add code file visitors and syntax rewriters for UblSharp generator
This commit introduces file and directory visitors along with syntax rewriters for property and field nodes. The document, project and workspace visitors are also added for enhanced file operations. The purpose is to simplify the codebase by replacing manual code manipulations with automatic Roslyn-based transformations. This consequently contributes to code readability and maintainability.